+++ flutter_extensions/lib/src/context.dart
@@ -2,23 +2,28 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 extension BuildContextDartExtensions on BuildContext {
+  /// The [ThemeData] instance from the closest context.
   ThemeData theme({bool shadowThemeOnly = false}) =>
       Theme.of(this, shadowThemeOnly: shadowThemeOnly);
 
+  /// The [DefaultTextStyle] instance from the closest context.
   DefaultTextStyle defaultTextStyle() => DefaultTextStyle.of(this);
 
+  /// The [IconThemeData] instance from the closest context.
   IconThemeData iconTheme() => IconTheme.of(this);
 
-  /// Localizations
-
+  /// The [Localizations] instance from the closest context.
   T localizations<T>() => Localizations.of<T>(this, T);
 
+  /// The [Locale] instance from the closest context.
   Locale locale({bool nullOk = false}) => Localizations.localeOf(this, nullOk: nullOk);
 
+  /// The [Navigator] instance from the closest context.
   NavigatorState navigator({bool rootNavigator = false, bool nullOk = false}) {
     return Navigator.of(this, rootNavigator: rootNavigator, nullOk: nullOk);
   }
 
+  /// Returns the read value if it exists otherwise null.
   T tryRead<T>() {
     try {
       return read<T>();
@@ -27,6 +32,7 @@ extension BuildContextDartExtensions on BuildContext {
     }
   }
 
+  /// Returns the watch value if it exists otherwise null.
   T tryWatch<T>() {
     try {
       return watch<T>();
@@ -37,6 +43,20 @@ extension BuildContextDartExtensions on BuildContext {
 }
 
 extension TargetPlatformDartExtensions on TargetPlatform {
-  bool get isMobile => this == TargetPlatform.android || this == TargetPlatform.iOS;
-  bool get isDesktop => this == TargetPlatform.fuchsia;
+  bool get isAndroid => this == TargetPlatform.android;
+  bool get isFuchsia => this == TargetPlatform.fuchsia;
+  bool get isIOS => this == TargetPlatform.iOS;
+  bool get isLinux => this == TargetPlatform.linux;
+  bool get isMacOS => this == TargetPlatform.macOS;
+  bool get isWindows => this == TargetPlatform.windows;
+
+  bool get isMobile {
+    switch (this) {
+      case TargetPlatform.android:
+      case TargetPlatform.iOS:
+        return true;
+      default:
+        return false;
+    }
+  }
 }