+++ pure_extensions/lib/built_extensions.dart
@@ -1,6 +1,8 @@
 library built_extensions;
 
-export 'src/built/GeoPointSerializer.dart';
-export 'src/built/async.dart';
+export 'src/built/by_pass_serializer.dart';
 export 'src/built/collections.dart';
+export 'src/built/geo_point_serializer.dart';
 export 'src/built/iterable.dart';
+export 'src/built/locale_serializer.dart';
+export 'src/built/serializers.dart';