+++ pure_extensions/lib/src/built/geo_point_serializer.dart
@@ -1,5 +1,5 @@
 import 'package:built_value/serializer.dart';
-import 'package:pure_extensions/src/dart/primitives/GeoPoint.dart';
+import 'package:pure_extensions/src/dart/primitives/geo_point.dart';
 
 /// Serializer for [GeoPoint] Type
 class GeoPointSerializer implements StructuredSerializer<GeoPoint> {