+++ pure_extensions/lib/src/dart/collections/iterable.dart
@@ -1,4 +1,8 @@
 import 'dart:async';
+import 'dart:math';
+
+import 'package:pure_extensions/src/dart/primitives/geo_point.dart';
+import 'package:rational/rational.dart';
 
 extension IterableExtensions<T> on Iterable<T> {
   /// returns null or if it is empty returns true.
@@ -95,7 +99,153 @@ extension IterableExtensions<T> on Iterable<T> {
   }
 }
 
+extension IterableNumDartExtension<T extends num> on Iterable<T> {
+  /// Calculate the sum of all numbers in the collection
+  T sumAll() {
+    if (isEmpty) return T is double ? 0.0 : 0;
+    return reduce((previousValue, element) => previousValue + element);
+  }
+
+  /// Calculate the subtraction of all numbers in the collection
+  T subtractAll() {
+    if (isEmpty) return T is double ? 0.0 : 0;
+    return reduce((previousValue, element) => previousValue - element);
+  }
+
+  /// Calculate the division of all numbers in the collection
+  double divideAll() {
+    return fold(0.0, (previousValue, element) => previousValue / element);
+  }
+
+  /// Calculate the multiplication of all numbers in the collection
+  T multiplyAll() {
+    if (isEmpty) return T is double ? 0.0 : 0;
+    return reduce((previousValue, element) => previousValue * element);
+  }
+
+  /// Sum [number] to all items in the collection
+  Iterable<T> sum(T number) {
+    return map((element) => element + number);
+  }
+
+  /// Subtract [number] to all items in the collection
+  Iterable<T> subtract(T number) {
+    return map((element) => element - number);
+  }
+
+  /// Divide [number] to all items in the collection
+  Iterable<double> divide(T number) {
+    return map((element) => element / number);
+  }
+
+  /// Multiply [number] to all items in the collection
+  Iterable<T> multiply(T number) {
+    return map((element) => element * number);
+  }
+
+  /// Calculate the average of all numbers in the collection
+  double average() => sumAll() / length;
+}
+
+extension IterableBigIntExtension<T extends BigInt> on Iterable<T> {
+  /// [IterableNumDartExtension.sumAll]
+  BigInt sumAll() {
+    if (isEmpty) return BigInt.zero;
+    return reduce((previousValue, element) => previousValue + element);
+  }
+
+  /// [IterableNumDartExtension.subtractAll]
+  BigInt subtractAll() {
+    if (isEmpty) return BigInt.zero;
+    return reduce((previousValue, element) => previousValue - element);
+  }
+
+  /// [IterableNumDartExtension.divideAll]
+  Rational divideAll() {
+    return fold(Rational.zero, (previousValue, element) => previousValue / Rational(element));
+  }
+
+  /// [IterableNumDartExtension.multiplyAll]
+  BigInt multiplyAll() {
+    if (isEmpty) return BigInt.zero;
+    return reduce((previousValue, element) => previousValue * element);
+  }
+
+  /// [IterableNumDartExtension.sum]
+  Iterable<T> sum(T number) {
+    return map((element) => element + number);
+  }
+
+  /// [IterableNumDartExtension.subtract]
+  Iterable<T> subtract(T number) {
+    return map((element) => element - number);
+  }
+
+  /// [IterableNumDartExtension.divide]
+  Iterable<Rational> divide(T number) {
+    final rationalNumber = Rational(number);
+    return map((element) => Rational(element) / rationalNumber);
+  }
+
+  /// [IterableNumDartExtension.multiply]
+  Iterable<T> multiply(T number) {
+    return map((element) => element * number);
+  }
+
+  /// [IterableNumDartExtension.average]
+  Rational average() => Rational(sumAll()) / Rational(BigInt.from(length));
+}
+
+extension IterableRationalExtension<T extends Rational> on Iterable<T> {
+  /// [IterableNumDartExtension.sumAll]
+  Rational sumAll() {
+    if (isEmpty) return Rational.zero;
+    return reduce((previousValue, element) => previousValue + element);
+  }
+
+  /// [IterableNumDartExtension.subtractAll]
+  Rational subtractAll() {
+    if (isEmpty) return Rational.zero;
+    return reduce((previousValue, element) => previousValue - element);
+  }
+
+  /// [IterableNumDartExtension.divideAll]
+  Rational divideAll() {
+    return fold(Rational.zero, (previousValue, element) => previousValue / element);
+  }
+
+  /// [IterableNumDartExtension.multiplyAll]
+  Rational multiplyAll() {
+    if (isEmpty) return Rational.zero;
+    return reduce((previousValue, element) => previousValue * element);
+  }
+
+  /// [IterableNumDartExtension.sum]
+  Iterable<T> sum(T number) {
+    return map((element) => element + number);
+  }
+
+  /// [IterableNumDartExtension.subtract]
+  Iterable<T> subtract(T number) {
+    return map((element) => element - number);
+  }
+
+  /// [IterableNumDartExtension.divide]
+  Iterable<Rational> divide(T number) {
+    return map((element) => element / number);
+  }
+
+  /// [IterableNumDartExtension.multiply]
+  Iterable<T> multiply(T number) {
+    return map((element) => element * number);
+  }
+
+  /// [IterableNumDartExtension.average]
+  Rational average() => sumAll() / Rational.fromInt(length);
+}
+
 extension IterableMapEntryExt<K, V> on Iterable<MapEntry<K, V>> {
+  /// Convert the collection of entries into a [Map].
   Map<K, V> toMap() => Map.fromEntries(this);
 
   /// Grouping the results in list according to key.
@@ -124,4 +274,44 @@ extension IterableExtFuture<T> on Iterable<Future<T>> {
 
   /// [Future.any]
   Future<T> anyFutures() => Future.any(this);
+}
+
+extension IterableGeoPointDartExtension on Iterable<GeoPoint> {
+  /// Calculate a center.
+  GeoPoint center() {
+    final eb = this.externalBounds();
+    final p = eb.northeast + eb.southwest;
+    return GeoPoint(p.latitude / 2, p.longitude / 2);
+  }
+
+  /// Calculate the northeast corner.
+  GeoPoint northeast() {
+    return reduce((p, c) {
+      return GeoPoint(max(p.latitude, c.latitude), max(p.longitude, c.longitude));
+    });
+  }
+
+  /// Calculate the southwest corner.
+  GeoPoint southwest() {
+    return reduce((p, c) {
+      return GeoPoint(min(p.latitude, c.latitude), min(p.longitude, c.longitude));
+    });
+  }
+
+  /// Calculate the internal corners.
+  GeoBounds internalBounds({GeoPoint center}) {
+    center ??= this.center();
+    return GeoBounds(
+      northeast: where((p) => p >= center).southwest(),
+      southwest: where((p) => p <= center).northeast(),
+    );
+  }
+
+  /// Calculate the external corners.
+  GeoBounds externalBounds() {
+    return GeoBounds(
+      northeast: northeast(),
+      southwest: southwest(),
+    );
+  }
 }