+++ pure_extensions/lib/src/dart/primitives/numbers.dart
@@ -6,6 +6,7 @@ extension NumDartExtensions on num {
   Rational toRational() => Rational.parse(toString());
 
   num min(num other) => math.min<num>(this, other);
+  num max(num other) => math.max<num>(this, other);
   num pow(num exponent) => math.pow(this, exponent);
   double sin() => math.sin(this);
   double cos() => math.cos(this);
@@ -16,6 +17,12 @@ extension NumDartExtensions on num {
   double sqrt() => math.sqrt(this);
   double exp() => math.exp(this);
   double log() => math.log(this);
+
+  /// Convert degrees to radians
+  num toRad() => this * (math.pi / 180.0);
+
+  /// Convert radians to degrees
+  num toDeg() => this * (180.0 / math.pi);
 }
 
 extension doubleDartExtensions on double {
@@ -23,6 +30,12 @@ extension doubleDartExtensions on double {
 
   double min(double other) => math.min<double>(this, other);
   double max(double other) => math.max<double>(this, other);
+
+  /// Convert degrees to radians
+  double toRad() => this * (math.pi / 180.0);
+
+  /// Convert radians to degrees
+  double toDeg() => this * (180.0 / math.pi);
 }
 
 extension intDartExtensions on int {