+++ pure_extensions/CHANGELOG.md
@@ -1,3 +1,8 @@
-## [0.0.1] - TODO: Add release date.
+## [2.0.2]
+- Export `GeoPoint`
 
-* TODO: Describe initial release.
+## [2.0.1]
+- Added `ByPassSerializer` and `LocaleSerializer`
+
+## [2.0.0]
+- Separate the dart language extensions from the flutter_extensions package